+++ tasty-plutus/CHANGELOG.md
@@ -4,6 +4,10 @@ This format is based on [Keep A Changelog](https://keepachangelog.com/en/1.0.0).
 
 ## Unreleased
 
+### Added
+
+* `passIf` combinator
+
 ### Changed
 
 * Rename `Example` to `Outcome` and its two constructors `Good` and `Bad` to